+++ app/Contracts/Services/Employee/EmployeeServiceInterface.php
@@ -11,12 +11,6 @@ use Illuminate\Http\Request;
 interface EmployeeServiceInterface
 {
     /**
-     * To get employee lists
-     * @return $array of employee
-     */
-    public function getEmployee();
-
-    /**
      * To search employee lists
      * 
      * @param Illuminate\Http\Request $request
@@ -30,9 +24,64 @@ interface EmployeeServiceInterface
      */
     public function showPieGraph();
 
-     /**
+    /**
      * To show bar graph
      * @return $array of employee
      */
     public function showBarGraph();
+
+    /**
+     * To get all employee list
+     * @param
+     * @return $employees
+     */
+    public function getAllEmployees();
+
+    /**
+     * To get employee lists
+     * @return $array of employee
+     */
+    public function getEmployee();
+
+    /**
+     * To get list of roles
+     *  @param
+     *  @return $roles
+     */
+    public function getRoles();
+
+    /**
+     * To get list of departs
+     *  @param
+     *  @return $departs
+     */
+    public function getDepartments();
+
+    /**
+     * To add new employee
+     * @param Request $request
+     * @return
+     */
+    public function addEmployee(Request $request);
+
+    /**
+     * To get a employee by id
+     * @param $id
+     * @return Object $employee
+     */
+    public function getEmployeeById($id);
+
+    /**
+     * To edit employee information
+     * @param $id,Request $request
+     * @return
+     */
+    public function editEmployeeById(Request $request, $id);
+
+    /**
+     * To delete employee by id
+     * @param $id
+     * @return
+     */
+    public function deleteEmployeeById($id);
 }
