+++ public/css/employee/employeecrud.css
@@ -1,0 +1,149 @@
+.cardcreate {
+  width: 70%;
+  margin: 0 auto;
+  margin-bottom: 30px;
+  padding-bottom: 30px;
+  box-shadow: 0 4px 8px 0 rgba(37, 18, 206, 0.2);
+}
+
+.card-header {
+  width: 100%;
+  padding: 20px;
+  background-color: blue;
+  color: #fff;
+  text-align: left;
+  font-size: 20px;
+}
+
+.card-body {
+  width: 60%;
+  margin-top: 10px;
+  margin-left: 20px;
+  text-align: left;
+}
+
+.con-left {
+  float: left;
+  width: 50%;
+  margin-bottom: 20px;
+}
+
+.con-right {
+  float: right;
+  margin-right: 13%;
+  /*height: 100px;*/
+  margin-bottom: 20px;
+  /*margin-left: 10%;*/
+  background-color: black;
+}
+
+.btn-cancel {
+  width: 100px;
+  padding-bottom: 10px;
+  background-color: red;
+  color: #fff;
+  font-size: 16px;
+  text-align: center;
+}
+
+.btn-back {
+  width: 100px;
+  margin-bottom: 20px;
+  padding-bottom: 10px;
+  background-color: rgb(250, 245, 245);
+  color: black;
+  font-size: 16px;
+  text-align: center;
+}
+
+.btn-add {
+  width: 130px;
+  padding-bottom: 10px;
+  background-color: blue;
+  color: #fff;
+  font-size: 16px;
+  text-align: center;
+}
+
+.form-label {
+  font-size: 16px;
+  font-weight: bold;
+}
+
+.form-control {
+  width: 500px;
+  margin-bottom: 20px;
+  padding: 10px;
+  outline: none;
+}
+
+.text-danger {
+  color: red;
+  font-weight: bold;
+}
+
+.card:hover {
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+}
+
+.container {
+  padding: 10px 16px;
+  color: #fff;
+  font-size: 20px;
+  text-align: center;
+}
+
+.title {
+  width: 500px;
+  margin: 0 auto;
+  color: black;
+  font-size: 30px;
+}
+
+select.form-select {
+  width: 500px;
+  margin-bottom: 20px;
+  margin-left: 20px;
+  padding: 0 10px;
+  padding-bottom: 20px;
+  background: #deece7;
+  color: #000;
+  font-size: 22px;
+}
+
+select.mySelect option {
+  width: 500px;
+  margin-bottom: 20px;
+  margin-left: 20px;
+  padding: 0 10px;
+  padding-bottom: 20px;
+  color: #000;
+}
+
+/*.modal fade{
+    width: 300px;
+    height: 200px;
+}
+.modal-dialog{
+    width: 300px;
+    height: 200px;
+    background-color: bisque;
+}
+.modal-header{
+    width: 300px;
+}
+.modal-title{
+    font-size: 30px;
+    color: black;
+}
+.modal-body{
+    text-align: center;
+    font-size: 20px;
+    color: red;
+    margin-top: 20px;
+    margin-bottom: 20px;
+}
+.modal-footer{
+width: 300px;
+background-color: aqua;
+}*/
