--- public/css/employeecrud.css
@@ -1,281 +1,0 @@
-/*.profile {
-    padding: 3px !important;
-}
-
-.title{
-    color: black;
-    font-size: 30px;
-}
-
-.container {
-    width: 800px;
-    margin: 30px auto;
-    padding: 50px;
-    color: black;
-  }
-
-  .profile-img {
-    width: 45px;
-    height: 45px;
-    margin-top: -1px;
-    margin-right: 16px;
-    border-radius: 50%;
-    border-color: blue;
-  }
-
-  .con {
-    width: 500px;
-    margin: 50px;
-    padding: 50px;
-    background-color: teal;
-    color:#fff;
-    text-align: center;
-    border-radius: 15px;
-  }
-  
-  h1 {
-    margin-bottom: 20px;
-    font-size: 35px;
-  }
-  
-  input {
-    margin: 20px 0px;
-  }
-  
-  input[type=text] {
-    width: 100%;
-    padding: 5px;
-    outline: none;
-  }
- 
-  input[type=submit] {
-    padding: 15px 25px;
-    border: 0;
-    background-color: #fff;
-    color: teal;
-  }
-  
-  img{
-    width: 200px;
-    height: 200px;
-  }
-
-  .btn{
-      width: 350px;
-      float: left;
-  }
-  .cancel{
-      width: 100px;
-      float: left;
-      color: #fff;
-      background-color: red;
-  }
-
-  .success{
-      width: 100px;
-      float: right;
-      color: #fff;
-      background-color: rgb(59, 39, 243);
-  }
-
-  .card {
-    width: 50%;
-    margin: 0 auto;
-    box-shadow: 0 4px 8px 0 rgba(37, 18, 206, 0.2);
-    transition: 0.3s;
-
-  }
-  .cardcreate {
-    width: 50%;
-    margin: 0 auto;
-    box-shadow: 0 4px 8px 0 rgba(37, 18, 206, 0.2);
-    transition: 0.3s;
-
-  }
-  .title{
-    width: 50%;
-    margin: 0 auto;
-    font-size: 30px;
-    font-weight: bold;
-  }
-  
-  .card:hover {
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-  }
-  
-  .container {
-    padding: 2px 16px;
-  }
-  */
-
-.profile-img {
-  width: 70%;
-  height: 350px;
-  margin-top: 10px;
-  margin-left: 5%;
-  text-align: center;
-  border-radius: 50%;
-  border-color: blue;
-}
-
-.card {
-  width: 40%;
-  margin: 0 auto;
-  margin-bottom: 30px;
-  box-shadow: 0 8px 8px 0 rgba(149, 221, 32, 0.2);
-  background-color: cadetblue;
-  transition: 0.3s;
-}
-
-.cardcreate {
-  width: 70%;
-  margin-bottom: 30px;
-  margin-left: 15%;
-  padding-bottom: 30px;
-  box-shadow: 0 4px 8px 0 rgba(37, 18, 206, 0.2);
-  transition: 0.3s;
-}
-
-.card-header {
-  width: 100%;
-  padding-top: 1px;
-  padding-bottom: 1px;
-  padding-left: 20px;
-  background-color: blue;
-  color: #fff;
-  text-align: left;
-}
-
-.card-body {
-  width: 60%;
-  margin-top: 10px;
-  margin-left: 20px;
-  text-align: left;
-}
-
-.con-left {
-  float: left;
-  width: 50%;
-  margin-bottom: 20px;
-}
-
-.con-right {
-  float: right;
-  margin-right: 13%;
-  /*height: 100px;*/
-  margin-bottom: 20px;
-  /*margin-left: 10%;*/
-  background-color: black;
-}
-
-.btn-cancel {
-  width: 100px;
-  padding-bottom: 10px;
-  background-color: red;
-  color: #fff;
-  font-size: 16px;
-  text-align: center;
-}
-
-.btn-back {
-  width: 100px;
-  margin-bottom: 20px;
-  padding-bottom: 10px;
-  background-color: rgb(250, 245, 245);
-  color: black;
-  font-size: 16px;
-  text-align: center;
-}
-
-.btn-add {
-  width: 130px;
-  padding-bottom: 10px;
-  background-color: blue;
-  color: #fff;
-  font-size: 16px;
-  text-align: center;
-}
-
-.form-label {
-  margin-left: 10px;
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.form-control {
-  width: 500px;
-  margin-bottom: 20px;
-  margin-left: 20px;
-  padding-bottom: 20px;
-}
-
-.text-danger {
-  color: red;
-  font-weight: bold;
-}
-
-.card:hover {
-  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
-}
-
-.container {
-  padding: 10px 16px;
-  color: #fff;
-  font-size: 20px;
-  text-align: center;
-}
-
-.title {
-  width: 500px;
-  margin: 0 auto;
-  color: black;
-  font-size: 30px;
-}
-
-select.form-select {
-  width: 500px;
-  margin-bottom: 20px;
-  margin-left: 20px;
-  padding: 0 10px;
-  padding-bottom: 20px;
-  background: #deece7;
-  color: #000;
-  font-size: 22px;
-}
-
-select.mySelect option {
-  width: 500px;
-  margin-bottom: 20px;
-  margin-left: 20px;
-  padding: 0 10px;
-  padding-bottom: 20px;
-  color: #000;
-}
-
-/*.modal fade{
-    width: 300px;
-    height: 200px;
-}
-.modal-dialog{
-    width: 300px;
-    height: 200px;
-    background-color: bisque;
-}
-.modal-header{
-    width: 300px;
-}
-.modal-title{
-    font-size: 30px;
-    color: black;
-}
-.modal-body{
-    text-align: center;
-    font-size: 20px;
-    color: red;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
-.modal-footer{
-width: 300px;
-background-color: aqua;
-}*/
